# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio backend capyba",
        "description": "Documentação da api desenvolvida para desafio backend para capyba - Gabriel Lima",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autorização"
                ],
                "description": "EP para dadastro de novo usuário no sistema.",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "photo"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha do usúario",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "Foto de perfil do funcionário",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autorização"
                ],
                "description": "EP para logar no sistema",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha do usúario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autorização"
                ],
                "description": "EP para deslogar no sistema",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/privacy/policy": {
            "get": {
                "tags": [
                    "Autorização"
                ],
                "description": "Termos de uso e politica de privacidade em PDF",
                "operationId": "d108ba8c8ebdd6b723dc8acf6091d53f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/code/confirm": {
            "post": {
                "tags": [
                    "Autorização"
                ],
                "description": "EP para confirmar codigo recebido por email",
                "operationId": "564695eeddb600c33244c43a89d7cbec",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Cádigo recebido por email",
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "code": 123456
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "codeAuth": []
                    }
                ]
            }
        },
        "/api/auth/code/resend": {
            "post": {
                "tags": [
                    "Autorização"
                ],
                "description": "EP para reenviar codigo de verificação por email",
                "operationId": "e270d8c83242306497ee15de43ab6a25",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email para reenviar codigo de ativação de conta",
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Para conseguir esse token, logue com email e senha com uma conta com email verificado",
                "name": "Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "codeAuth": {
                "type": "http",
                "description": "Para conseguir esse token, crie uma conta, ou envie novamente um codigo para email de uma conta já cadastrada, que esse token virá na resposta da requisição",
                "name": "Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}